각 빌더 root 레벨로 분리 및 order빌더와 where or,and 메소드 추가
