+++ src/main/java/prgms/spring_week1/domain/customer/repository/impl/JdbcCustomerRepository.java
@@ -8,7 +8,10 @@ import org.springframework.stereotype.Component;
 import prgms.spring_week1.domain.customer.model.Customer;
 import prgms.spring_week1.domain.customer.model.embeddedType.Email;
 import prgms.spring_week1.domain.customer.repository.CustomerRepository;
-import prgms.spring_week1.domain.util.SqlBuilder;
+import prgms.spring_week1.domain.util.sqlBuilder.actionBuilder.DeleteBuilder;
+import prgms.spring_week1.domain.util.sqlBuilder.actionBuilder.InsertBuilder;
+import prgms.spring_week1.domain.util.sqlBuilder.actionBuilder.SelectBuilder;
+import prgms.spring_week1.domain.util.sqlBuilder.actionBuilder.UpdateBuilder;
 import prgms.spring_week1.domain.voucher.repository.impl.JdbcVoucherRepository;
 
 import javax.sql.DataSource;
@@ -16,7 +19,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 @Component
 public class JdbcCustomerRepository implements CustomerRepository {
@@ -51,7 +53,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
 
     @Override
     public void insert(Customer customer) {
-        String insertSQL = new SqlBuilder.InsertBuilder()
+        String insertSQL = new InsertBuilder()
                 .insert("customers")
                 .columns("customer_id","email", "name")
                 .values("UUID_TO_BIN(:customerId)",":email",":name")
@@ -63,7 +65,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
     @Override
     public List<Customer> findAll() {
 
-        String findAllSql = new SqlBuilder.SelectBuilder()
+        String findAllSql = new SelectBuilder()
                     .select("*")
                     .from("customers")
                     .build();
@@ -73,7 +75,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
 
     @Override
     public Customer findByEmail(String email) {
-        String findByEmailSql = new SqlBuilder.SelectBuilder()
+        String findByEmailSql = new SelectBuilder()
                 .select("*")
                 .from("customers")
                 .where("email = :email")
@@ -90,7 +92,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
 
     @Override
     public void updateInfo(String beforeUpdateEmail, String afterUpdateEmail) {
-        String updateInfoSql = new SqlBuilder.UpdateBuilder()
+        String updateInfoSql = new UpdateBuilder()
                 .update("customers")
                 .set("email = :afterUpdateEmail")
                 .where("email = :beforeUpdateEmail")
@@ -101,7 +103,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
 
     @Override
     public void deleteByEmail(String email) {
-        String deleteByEmailSql = new SqlBuilder.DeleteBuilder()
+        String deleteByEmailSql = new DeleteBuilder()
                 .delete()
                 .from("customers")
                 .where("email = :email")
@@ -112,7 +114,7 @@ public class JdbcCustomerRepository implements CustomerRepository {
 
     @Override
     public void deleteAll() {
-        String deleteAllSql = new SqlBuilder.DeleteBuilder()
+        String deleteAllSql = new DeleteBuilder()
                 .delete()
                 .from("customers")
                 .build();
