+++ src/main/java/prgms/spring_week1/domain/util/sqlBuilder/LastIndexLength.java
@@ -1,0 +1,6 @@
+package prgms.spring_week1.domain.util.sqlBuilder;
+
+public class LastIndexLength {
+    public static int INVALID_LAST_COMMA = 1;
+    public static int INVALID_LAST_AND = 5;
+}
