+++ src/main/java/prgms/spring_week1/domain/util/sqlBuilder/tableBuilder/ValuesBuilder.java
@@ -1,0 +1,13 @@
+package prgms.spring_week1.domain.util.sqlBuilder.tableBuilder;
+
+public class ValuesBuilder {
+    private StringBuilder valuesBuilder;
+
+    public ValuesBuilder(StringBuilder valuesBuilder) {
+        this.valuesBuilder = valuesBuilder;
+    }
+
+    public String build(){
+        return valuesBuilder.toString();
+    }
+}
